\name{Spgr-package}
\alias{Spgr-package}
\docType{package}
\title{Spgr}
\description{Spgr provides functions to estimate subgroups in linear regression via concave fusion based on SCAD penalty with repeated measures. Weights associated with pairwise penalties can be specified.}
\details{
 Spgr can be used to estimate subgroups based on liear regression coefficients via concave fusion penalties. Both individial data and repeated data can be applied. It allows to fit spatial data or non-spatial data. For fitting spatial data, weights can be specified. Two different BICs forms are provided. 
}
\author{
  Xin Wang
  
  Maintainer: Xin Wang <wangxin8923@gmail.com>
}
\references{
  Ma, S. and Huang, J. (2018). Exploration of heterogeneous treatment effects via concave fusion. arXiv preprint arXiv:1607.03717.
  
Ma, S. and Huang, J. (2017). A concave pairwise fusion approach to subgroup analysis. Journal of the American Statistical Association, 112(517):410–423.

Wang, X., Zhu, Z and Zhang, H. (2019). Estimating subgroups with spatial information with repeated measures. 
}

